A lot of new stuff.
Handles are nearly done.  At that point, I need only introduce cell decompositions and dual cell structures to the triangulations section.

The discussion in section 3.1 offers a much more concise argument to support the method of building 4-manifolds.  The current plan is to remove any mention of shadows to reduce the burden placed on language.

The algorithms of chapters 4/5/6 will be revised to a new format.  See the incomplete section 5.3 for an example.  The structure will be an introduction of data types used followed by a brief callback to the smooth case, followed finally by a precise method.

Signed-off-by: smchurchill <meadowschurchill@gmail.com>